part2/requirements.txt
Flask==2.2.3
flask-restx==1.1.0
flask-cors==3.0.10
PyJWT==2.6.0
werkzeug==2.2.3
gunicorn==20.1.0
pytest==7.3.0
python-dotenv==1.0.0